change vimrc foldmarker to [,]
Previously it was {,} and this was causing problems with folding.
Specificially, UltiSnips uses a ${n} notation for tabstops. But, with
foldmarker set to "{,}", a "{n" implies the foldlevel. Particularly,
"{1" (which is used quite often for UltiSnips) sets the foldlevel to 1
an destroys all folds. Changing to "[,]" solves this, but some extra
care has to be taken that what vim thinks are folds are always matched.
For example, consider a map like
 	noremap . .`[
This leaves a dangling braacket; you must add an extra "] below it to
"close" the fold you just introduced.

I guess this is why the vim docs suggest using the ugly (but unique)
"{{{,}}}".